# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_thepantheon_api_internal_model.Course:
    properties:
      categoria:
        $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseCategory'
      categoria_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      imagem:
        type: string
      modulos:
        items:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseModule'
        type: array
      nome:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_thepantheon_api_internal_model.CourseCategory:
    properties:
      created_at:
        type: string
      cursos:
        items:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.Course'
        type: array
      id:
        type: string
      imagem:
        type: string
      nome:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_thepantheon_api_internal_model.CourseItem:
    properties:
      conteudo:
        type: string
      created_at:
        type: string
      id:
        type: string
      modulo_id:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_thepantheon_api_internal_model.CourseModule:
    properties:
      created_at:
        type: string
      curso_id:
        type: string
      id:
        type: string
      modulo:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.CreateCourseCategoryRequest:
    properties:
      cursos_ids:
        items:
          type: string
        type: array
      imagem:
        type: string
      nome:
        minLength: 2
        type: string
    required:
    - nome
    type: object
  github_com_thepantheon_api_internal_model.CreateCourseItemRequest:
    properties:
      conteudo:
        minLength: 2
        type: string
      modulo_id:
        type: string
      tipo:
        minLength: 2
        type: string
      titulo:
        minLength: 2
        type: string
    required:
    - conteudo
    - tipo
    - titulo
    type: object
  github_com_thepantheon_api_internal_model.CreateCourseModuleRequest:
    properties:
      itens_ids:
        items:
          type: string
        type: array
      modulo:
        minLength: 2
        type: string
    required:
    - modulo
    type: object
  github_com_thepantheon_api_internal_model.CreateCourseRequest:
    properties:
      categoria_id:
        type: string
      imagem:
        type: string
      modulos_ids:
        items:
          type: string
        type: array
      nome:
        minLength: 2
        type: string
    required:
    - nome
    type: object
...
  github_com_thepantheon_api_internal_model.UpdateCourseCategoryRequest:
    properties:
      cursos_ids:
        items:
          type: string
        type: array
      imagem:
        type: string
      nome:
        minLength: 2
        type: string
    type: object
  github_com_thepantheon_api_internal_model.UpdateCourseItemRequest:
    properties:
      conteudo:
        minLength: 2
        type: string
      modulo_id:
        type: string
      tipo:
        minLength: 2
        type: string
      titulo:
        minLength: 2
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.UpdateCourseRequest:
    properties:
      categoria_id:
        type: string
      imagem:
        type: string
      modulos_ids:
        items:
          type: string
        type: array
      nome:
        minLength: 2
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pantheonconcursos.com.br
    name: API Support
  description: API REST para sistema de concursos com autenticação JWT
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ThePantheon API
  version: "1.0"
paths:
...
  /cursos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.Course'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar cursos
      tags:
      - cursos
    post:
      consumes:
      - application/json
      parameters:
      - description: Curso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.Course'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar curso
      tags:
      - cursos
  /cursos/{id}:
    delete:
      parameters:
      - description: ID do curso
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover curso
      tags:
      - cursos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do curso
        in: path
        name: id
        required: true
        type: string
      - description: Curso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.Course'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar curso
      tags:
      - cursos
  /cursos/categorias:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseCategory'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar categorias
      tags:
      - categorias
    post:
      consumes:
      - application/json
      parameters:
      - description: Categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateCourseCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar categoria
      tags:
      - categorias
  /cursos/categorias/{id}:
    delete:
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover categoria
      tags:
      - categorias
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateCourseCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar categoria
      tags:
      - categorias
...
  /meus-cursos/itens:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar itens
      tags:
      - meus-cursos
    post:
      consumes:
      - application/json
      parameters:
      - description: Item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateCourseItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar item sem modulo
      tags:
      - meus-cursos
  /meus-cursos/itens/{id}:
    delete:
      parameters:
      - description: ID do item
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover item
      tags:
      - meus-cursos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do item
        in: path
        name: id
        required: true
        type: string
      - description: Item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateCourseItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar item
      tags:
      - meus-cursos
  /meus-cursos/modulos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseModule'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar modulos
      tags:
      - meus-cursos
    post:
      consumes:
      - application/json
      parameters:
      - description: Modulo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateCourseModuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseModule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar modulo
      tags:
      - meus-cursos
  /meus-cursos/modulos/{id}:
    delete:
      parameters:
      - description: ID do modulo
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover modulo
      tags:
      - meus-cursos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do modulo
        in: path
        name: id
        required: true
        type: string
      - description: Modulo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateCourseModuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CourseModule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar modulo
      tags:
      - meus-cursos
...
securityDefinitions:
  Bearer:
    description: Digite "Bearer" seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"